**Before**

```
Benchmarking: anytime_beam_search(board, {'power': 10})
 ..........
 Solutions of avg. length 15.500 in 19.303 sec.
 Avg. length times sec (lower is better) => 299.204

Benchmarking: heuristic_search(board, {'max_nodes': 220000})
 ..........
 Solutions of avg. length 17.600 in 18.903 sec.
 Avg. length times sec (lower is better) => 332.688

Benchmarking: monte_carlo_search(board, {'iterations': 3100, 'seed': 1})
 ..........
 Solutions of avg. length 16.500 in 20.085 sec.
 Avg. length times sec (lower is better) => 331.403
```

**After**

```
Benchmarking: anytime_beam_search(board, {'power': 10})
 ..........
 Solutions of avg. length 15.500 in 13.073 sec.
 Avg. length times sec (lower is better) => 202.630

Benchmarking: heuristic_search(board, {'max_nodes': 220000})
 ..........
 Solutions of avg. length 17.600 in 14.685 sec.
 Avg. length times sec (lower is better) => 258.453

Benchmarking: monte_carlo_search(board, {'iterations': 3100, 'seed': 1})
 ..........
 Solutions of avg. length 16.500 in 14.271 sec.
 Avg. length times sec (lower is better) => 235.474
```